requirements.txt
tensorflow
fpdf
PyYAML
streamlit
streamlit_navigation_bar
transformers
youtube_transcript_api


